+++ nall/utility.hpp
@@ -20,9 +20,9 @@ template<typename To, typename With> struct castable {
   With value;
 };
 
-template<typename T> inline auto allocate(uint size, const T& value) -> T* {
+template<typename T> inline auto allocate(u64 size, const T& value) -> T* {
   T* array = new T[size];
-  for(uint i = 0; i < size; i++) array[i] = value;
+  for(u64 i = 0; i < size; i++) array[i] = value;
   return array;
 }
